Use uint64 instead of int64 in QGB types (#358)
* uses the uint64 instead of int64 in right places

* typo

* cosmetics

+++ x/qgb/handler_test.go
@@ -113,7 +113,7 @@ func TestMsgDataCommitmentConfirm(t *testing.T) {
 	require.NoError(t, err)
 	dataHash := types.DataCommitmentTupleRootSignBytes(
 		types.BridgeId,
-		big.NewInt(100/types.DataCommitmentWindow),
+		big.NewInt(int64(100/types.DataCommitmentWindow)),
 		bytesCommitment,
 	)
